fix: restore ::float8 casts and set microdollars_balance in test helper
The ::float8 cast is required because PostgreSQL bigint subtraction
returns bigint, and the pg driver maps computed bigint expressions to
JS BigInt (not number). Without the cast, arithmetic with JS numbers
throws 'Cannot mix BigInt and other types'.

Also set microdollars_balance in createTestOrganization helper to keep
the deprecated column consistent for any tests that read it.

+++ src/app/api/cron/oss-monthly-reset/route.ts
@@ -82,7 +82,7 @@ export async function GET(request: Request) {
       id: organizations.id,
       name: organizations.name,
       currentBalance:
-        sql<number>`(${organizations.total_microdollars_acquired} - ${organizations.microdollars_used})`.as(
+        sql<number>`(${organizations.total_microdollars_acquired} - ${organizations.microdollars_used})::float8`.as(
           'currentBalance'
         ),
       settings: organizations.settings,
